requirements.txt
torch>=2.2
timm>=0.9.16
numpy>=1.24
pillow>=10.0
opencv-python>=4.8
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
pyyaml>=6.0
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
python-multipart>=0.0.9

# Hugging Face dataset downloader
datasets>=2.18

# Explainability
# Optional Grad-CAM library (import name `pytorch_grad_cam`). Fallback implementation included.
# grad-cam>=1.5.0

# Agentic orchestration
langgraph>=0.0.55

# Optional LLM client (OpenAI-compatible)
openai>=1.12
httpx>=0.26

# UI dashboard
streamlit>=1.32
pandas>=2.1
requests>=2.31
